+++ Scripts/Cores/BaseDialoguePreProcessing.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace DSC.DialogueSystem
+{
+    public abstract class BaseDialoguePreProcessing : MonoBehaviour
+    {
+        public abstract void PreProcessingDialogue(ref Dialogue hDialogue); 
+    }
+}